+++ src/test/java/org/bytedeco/javacpp/PointerTest.java
@@ -958,12 +958,13 @@ public class PointerTest {
     @Test public void testPointerScope() {
         System.out.println("PointerScope");
         IntPointer outside = new IntPointer(1);
-        IntPointer attached = new IntPointer(1), detached, inside, inside1, inside2, inside3, inside4, inside5;
+        IntPointer attached = new IntPointer(1), detached, inside, inside1, inside2, retained1, retained2, inside5;
 
         try (PointerScope scope = new PointerScope()) {
             scope.attach(attached);
 
             detached = new IntPointer(1);
+            detached.retainReference();
             scope.detach(detached);
 
             inside = new IntPointer(1);
@@ -971,10 +972,13 @@ public class PointerTest {
                 inside1 = new IntPointer(1);
                 inside2 = new IntPointer(1);
             }
-            try (PointerScope scope2 = new PointerScope(false)) {
-                inside3 = new IntPointer(1);
-                inside4 = new IntPointer(1);
+            try (PointerScope scope2 = new PointerScope()) {
+                retained1 = new IntPointer(1);
+                retained2 = new IntPointer(1);
+                retained1.retainReference();
+                scope.attach(retained2);
             }
+            retained2.retainReference();
             inside5 = new IntPointer(1);
         }
 
@@ -986,15 +990,27 @@ public class PointerTest {
         assertTrue(inside.isNull());
         assertTrue(inside1.isNull());
         assertTrue(inside2.isNull());
-        assertFalse(inside3.isNull());
-        assertFalse(inside4.isNull());
+        assertFalse(retained1.isNull());
+        assertFalse(retained2.isNull());
         assertTrue(inside5.isNull());
         assertFalse(outside2.isNull());
+
+        outside.releaseReference();
+        detached.releaseReference();
+        retained1.releaseReference();
+        retained2.releaseReference();
+        outside2.releaseReference();
+
+        assertTrue(outside.isNull());
+        assertTrue(detached.isNull());
+        assertTrue(retained1.isNull());
+        assertTrue(retained2.isNull());
+        assertTrue(outside2.isNull());
 
         IntPointer intPointer;
         FloatPointer floatPointer;
         try (PointerScope globalScope = new PointerScope()) {
-            try (PointerScope localScope = new PointerScope(true, IntPointer.class)) {
+            try (PointerScope localScope = new PointerScope(IntPointer.class)) {
                 intPointer = new IntPointer(1);
                 floatPointer = new FloatPointer(1);